requirements.txt
# Dev environment
pip
autopep8

# App
flask
google-cloud-texttospeech
python-dotenv
gunicorn
flask-cors

